Log latest parsed event instead of sample parsed event
If honeytail gets "stuck" tailing a large / high-volume file, it can be tricky
to figure out exactly where it got stuck. In order to make that a little
easier, print the "last parsed event" with every status summary.

This replaces the "sample parsed event", which was the first
event read off the responses queue within that status interval.