Add Postgres slow query parser
This is a fairly minimal stab at slurping Postgres query logs. In Postgres,
slow queries go into the general log if you set the
`log_min_duration_statement` setting.

There are a number of things that I've omitted from this patch for the moment,
in the interest of simplicity:

- Running multiple parser goroutines in parallel.
- Propagating a sample rate down to the parser.
- Verifying compatibility across postgres versions -- I've only tested this
  with PostgreSQL 9.5.
- Parsing log statements that are not slow queries.

I plan to address the first three in subsequent patches, but might punt on
parsing other types of log statements unless there's demonstrated interest.