adding data augmentation map feature (#98)
This change adds a feature to honeytail - you can hand it a map of keys to values and it will inspect each event for those keys in a specific column. If the key is found, it adds all fields in the value to the event.

There are several use cases for this when parsing logs:

    the audit log for RDS only contains IP addresses. It would be nice to add hostnames (and host groups!) to the events sent in to Honeycomb to make grouping easier "Is this db request from an app server?"
    ssh connections log the fingerprint of the key used to authenticate - it would be nice to translate key fingerprints to usernames to show more easily who it is that's connecting to a server
    the log contains applicaiton IDs and you want to include application name, pricing tier, and db shard to the event.

The usage pattern for any of these cases would be to generate a map file that goes from the keys (ip address, ssh key fingerprint, appID, etc.) to a list of key/value pairs that you want to include. Distribute this map file to your servers running honeytail and configure it to take action based on the contents of the map file.