Add flatten capability to json parser
Some apps generate valuable json output but the meat of it is one layer down, wrapped by some metadata. It would be helpful to ingest this data without writing a custom parser for it. This adds --depth option to the json parser, allowing it to flatten down to the specified level, merging json keys with a delimiter. Used incorrectly, this could cause an explosion in columns, but I'd rather have this capability than not.