Add --rebase_time argument
When set, timestamps are "rebased" onto the current time. The last event (chronologically) in the log file(s) is set to the current time. Event timestamps before this are updated relative to this time, with the time delta preserved.

For example, if now is Wed Jul 4 12:00:00 -0800 PST 2018, and the events start and end at Wed Jul 3 15:04:05 -0800 PST 2018 and Wed Jul 3 12:00:05 -0800 PST 2018, respectively, the events will be updated so that they appear to have started and ended at Wed Jul 4 08:56:00 -0800 PST 2018 and Wed Jul 4 12:00:00 -0800 PST 2018.

The goal is to allow you to input old data but look at it as though it happened recently.