WIP: to be rebased. Actually works on apache logs now.
In the Test code we're using Json. Json just assumes every numeric is a float.
However, in the magical Nginx parsing code, it will only cast to a float if the
number has a decimal place. Otherwise it's an int.

This code is longer than I want and unsure if everything passing through will be
an int64 or a float64.