Update Go to 1.16 & test cross-compilation in PRs (#200)
* test and build with Go 1.16

Not only is it the latest and greatest, it allows for darwin-arm64 builds.

* matrixify the 'go build' for os/arch

Turn the parameterized 'go-build' command into a parameterized
'go-build' job. As a job separate from the packaging, this can be added
to the 'build' workflow as a job to perform on pull requests and not
only run on git taggings.

* build directly to ~/artifacts/ in CI

This simplifies the hand-offs of the built binaries to subsequent jobs
in a workflow.

* remove Windows from the build matrix

The amd64 version throws errors during the build:

    # github.com/honeycombio/honeytail/tail
    tail/tail.go:321:13: undefined: unix.Stat_t
    tail/tail.go:322:12: undefined: unix.Stat
    tail/tail.go:431:13: undefined: unix.Stat_t
    tail/tail.go:432:2: undefined: unix.Stat
    Exited with code exit status 2

Probably need to use os.Stat() instead?

Also, need to figure out how to exclude the Windows + arm64 combination
from the matrix. If we can get 'exclude:' to work here, we can add a
conditional to the job[1] to do nothing.

[1] https://circleci.com/docs/2.0/reusing-config/#defining-conditional-steps