feat: Add config option for log level (#332)
## Which problem is this PR solving?
Honeytail logs a summary of every batch of events sent to Honeycomb.
These are very repetitive and obscure other information if there was a
problem sending events. The summaries are logged at the `INFO` level
using logrus.

This PR adds a log level configuration option that allows users to set
their desired log level in the logging framework logrus. The usage of
info level was verified to not contain critical information if there was
a problem, with more important logs using a higher log level (eg
`error`).

- Closes #330 

## Short description of the changes
- Add new `log_level` option, defaulting to `info`
- If the new option has a value, attempt to set that log level in logrus

This was tested by setting the `log_level` flag after building the tool
locally. As we use flag parsing in main file, it's not easy to write
unit tests.